requirements.txt
allure-pytest
allure-python-commons
Appium-Python-Client
configparser
faker
flaky
gevent
oauthlib
parameterized
prospector
py
pytest
pytest-repeat
pytest-rerunfailures
pytest-xdist
requests
requests_oauthlib
xmltodict
